+++ generator/src/jvmMain/kotlin/tech/skot/tools/generation/SuppressWarningsNames.kt
@@ -1,0 +1,6 @@
+package tech.skot.tools.generation
+
+object SuppressWarningsNames {
+    val resourcesWarning = listOf("PropertyName")
+    val genWarning = listOf("REDUNDANT_PROJECTION","RedundantVisibilityModifier")
+}