+++ plugin/src/main/kotlin/tech/skot/Versions.kt
@@ -7,21 +7,21 @@ import kotlin.String
  * This code is generated from build.gradle.kts
  */
 public object Versions {
-  public const val skot: String = "1.3.9-ua"
+  public const val skot: String = "1.3.10-ua"
 
   public const val group: String = "com.github.useradgents.skot"
 
-  public const val serialization: String = "1.8.0"
+  public const val serialization: String = "1.8.1"
 
-  public const val kotlinxDateTime: String = "0.6.1"
+  public const val kotlinxDateTime: String = "0.6.2"
 
   public const val ktor: String = "3.1.2"
 
   public const val kotlin: String = "2.1.20"
 
-  public const val kotlinCoroutines: String = "1.10.1"
+  public const val kotlinCoroutines: String = "1.10.2"
 
-  public const val kotlinpoet: String = "1.18.1"
+  public const val kotlinpoet: String = "2.1.0"
 
   public const val android_minSdk: Int = 23
 
@@ -33,9 +33,9 @@ public object Versions {
 
   public const val sqldelight: String = "1.5.5"
 
-  public const val desugaring: String = "2.1.4"
+  public const val desugaring: String = "2.1.5"
 
   public const val leakcanary: String = "2.14"
 
-  public const val ktlint: String = "1.3.1"
+  public const val ktlint: String = "1.5.0"
 }